Update meter_page : Seperated meter_control_button widget
