+++ lib/pages/meter_page/meter_page.dart
@@ -336,47 +336,3 @@ class _MeterInfoState extends State<MeterInfo> {
     );
   }
 }
-
-class MeterButton extends StatefulWidget {
-  const MeterButton({super.key, required this.btnColor, required this.btnText, required this.onClickFunction});
-
-  final Color btnColor;
-  final String btnText;
-  final void Function() onClickFunction;
-
-  @override
-  State<StatefulWidget> createState() => _MeterButtonState();
-}
-
-class _MeterButtonState extends State<MeterButton> {
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 5.0),
-        child: Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(100),
-            color: widget.btnColor
-          ),
-          child: InkWell(
-            onTap: widget.onClickFunction,
-            child: Center(
-              child: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 20.0),
-                child: Text(
-                  widget.btnText,
-                  textAlign: TextAlign.center,
-                  style: const TextStyle(
-                    color: MeterColor.meterBtnText,
-                    fontSize: 17.5
-                  ),
-                ),
-              )
-            ),
-          )
-        ),
-      )
-    );
-  }
-}
